Reorganize remaining base classes.
Move Estimator and Transformer into base.py.
Move base/base.py up a level to just base.py.
Move MetaResult into its own file at the top level (results.py).
