+++ nimare/meta/cbma/mkda.py
@@ -12,7 +12,7 @@ from nilearn.masking import apply_mask, unmask
 from statsmodels.sandbox.stats.multicomp import multipletests
 
 from .kernel import MKDAKernel, KDAKernel
-from ...base import MetaResult
+from ...results import MetaResult
 from .base import CBMAEstimator
 from .kernel import KernelTransformer
 from ...stats import null_to_p, p_to_z, one_way, two_way
